Eliminate Codeception\TestRunner class
